requirements.txt
python-dotenv==1.0.1
aiogram==2.9.2
